backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pillow==10.0.1
torch==2.1.0
torchvision==0.16.0
opencv-python==4.8.1.78
numpy==1.24.3
requests==2.31.0
gdown==4.7.1
werkzeug==2.3.7